+++ zebra-consensus/src/error.rs
@@ -12,7 +12,7 @@ use thiserror::Error;
 
 use zcash_protocol::value::BalanceError;
 use zebra_chain::{
-    amount, block, orchard,
+    amount, block, ironwood, orchard,
     parameters::subsidy::SubsidyError,
     sapling, sprout,
     transparent::{self, MIN_TRANSPARENT_COINBASE_MATURITY},
@@ -49,6 +49,9 @@ pub enum TransactionError {
 
     #[error("coinbase transaction MUST NOT have the EnableSpendsOrchard flag set")]
     CoinbaseHasEnableSpendsOrchard,
+
+    #[error("coinbase transaction MUST NOT have the EnableSpendsIronwood flag set")]
+    CoinbaseHasEnableSpendsIronwood,
 
     #[error("coinbase transaction MUST NOT have an Orchard shielded bundle")]
     CoinbaseHasOrchardShieldedData,
@@ -183,8 +186,17 @@ pub enum TransactionError {
     #[error("orchard double-spend: duplicate nullifier: {_0:?}")]
     DuplicateOrchardNullifier(orchard::Nullifier),
 
+    #[error("ironwood double-spend: duplicate nullifier: {_0:?}")]
+    DuplicateIronwoodNullifier(ironwood::Nullifier),
+
     #[error("must have at least one active orchard flag")]
     NotEnoughFlags,
+
+    #[error("must have at least enable spend or enable output flag set")]
+    NotEnoughIronwoodFlags,
+
+    #[error("Orchard transactions MUST NOT have the EnableCrossAddress flag set")]
+    OrchardHasEnableCrossAddress,
 
     #[error("could not find transparent input UTXO in the best chain or mempool")]
     TransparentInputNotFound,
@@ -250,6 +262,9 @@ pub enum TransactionError {
 
     #[error("Orchard proof has a non-canonical size")]
     OrchardProofSize,
+
+    #[error("Ironwood proof has a non-canonical size")]
+    IronwoodProofSize,
 
     #[error("unexpected error")]
     Other(String),
@@ -345,6 +360,7 @@ impl TransactionError {
             | CoinbaseHasSpend
             | CoinbaseHasOutputPreHeartwood
             | CoinbaseHasEnableSpendsOrchard
+            | CoinbaseHasEnableSpendsIronwood
             | CoinbaseHasOrchardShieldedData
             | CoinbaseOutputsNotDecryptable
             | CoinbaseInMempool
@@ -367,6 +383,8 @@ impl TransactionError {
             | DisabledAddToSproutPool
             | DisabledAddToOrchardPool
             | NotEnoughFlags
+            | NotEnoughIronwoodFlags
+            | OrchardHasEnableCrossAddress
             | WrongConsensusBranchId
             | MissingConsensusBranchId
             | LockedUntilAfterBlockHeight(_)